btrfs-progs: cast u64 to long long to avoid printf warnings
When building on ppc64 I hit a number of warnings in printf:

btrfs-map-logical.c:69: error: format ‘%Lu’ expects type ‘long long
unsigned int’, but argument 4 has type ‘u64’

Fix them.

+++ btrfs-list.c
@@ -249,7 +249,8 @@ static int resolve_root(struct root_lookup *rl, struct root_info *ri)
 			break;
 		}
 	}
-	printf("ID %llu top level %llu path %s\n", ri->root_id, top_id,
+	printf("ID %llu top level %llu path %s\n",
+	       (unsigned long long)ri->root_id, (unsigned long long)top_id,
 	       full_path);
 	free(full_path);
 	return 0;
